requirements.txt
torch
torchvision
streamlit
numpy
pillow
